+++ resources/js/components/Cart.vue
@@ -95,7 +95,7 @@
                 } else {
                     // assegna indirizzo di spedizione di default
                     //
-                    axios.get('/api/billing-profiles/' + event.id + '/shipping-address').then(response => {
+                    axios.get('/billing-profiles/' + event.id + '/shipping-address').then(response => {
                         this.shipping_address = response.data;
                     });
                 }